share retry code across both implementations
This also changes the retry implementation and allows us to send
files as large as the size_t limit on the machine with a single
method dispatch.

While we're at it, rename our retrying function to "sendfile_full"
(from "__sendfile"), as functions prefixed with "__" are
reserved for the compiler and standard library internals.