avoid hogging the 1.8 interpreter with large files
With the non-native threads of Ruby 1.8, sendfile() with a large
file can block other threads indefinitely if the O_NONBLOCK flag
is unset on the socket.  So only run sendfile() on 64K blocks
and allow other threads to run in-between sendfile() calls.

64K was chosen because it is half the typical maximum readahead
size of Linux 2.6, giving the kernel some time to populate the
page cache in between sendfile() calls if other threads are
running.